Cleanup File Directories
Moved scripts into categorical folders.
