+++ Global Game Jam 2024/Assets/Scripts/Clown.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aa120ba2ed89cbf44ad57e05a14a0147
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
