+++ Global Game Jam 2024/Assets/Scripts/Clown/SpawnClown.cs
@@ -1,0 +1,15 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class SpawnClown : MonoBehaviour
+{
+    [SerializeField] GameObject clown;
+    public void spawnClown()
+    {
+        LevelController.Instance.livesLeft--;
+        GameObject newClown = Instantiate(clown);
+        Vector2 Position = this.transform.position;
+        newClown.transform.position = Position;
+    }
+}
