+++ Global Game Jam 2024/Assets/Scripts/Obstacle.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0f78a9858b9b8a044a1d6cca268a7ba1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
