+++ Global Game Jam 2024/Assets/Scripts/Scene.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 379931a61cf49f34ea10c07f37109083
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
