+++ Global Game Jam 2024/Assets/Scripts/Scene/Parallax.cs
