+++ Global Game Jam 2024/Assets/Scripts/Scene/ParallaxFactory.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a9c7c01ec63c1ed4ab25b78abacc8de7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
